Refactor batchnorm + graph module.
This commit
- Removes graph.py and splits it across an __init__.py and several
  other .py files under graph/ (beginning the process accomplished
  for bricks/ in the previous commit).
- Splits the batch normalization implementation across bricks/ and
  graph/ in the expected fashion (graph transformation function
  under graph/, bricks under bricks/).
- Makes numerous small improvements to the batch normalization
  implementation.
