+++ blocks/bricks/__init__.py
@@ -1,5 +1,7 @@
 """Bricks are parameterized Theano operations."""
 from .base import application, Brick, lazy
+from .bn import (BatchNormalization, SpatialBatchNormalization,
+                 BatchNormalizedMLP)
 from .interfaces import Activation, Feedforward, Initializable, Random
 from .simple import (Linear, Bias, Maxout, LinearMaxout, Identity, Tanh,
                      Logistic, Softplus, Rectifier, Softmax,
