+++ blocks/bricks/bn.py
@@ -1,23 +1,23 @@
 import collections
+
 import numpy
-from theano import tensor
-from theano.tensor.nnet import bn
-from blocks.bricks import (RNGMixin, lazy, application, Sequence,
-                           Feedforward, MLP)
-from blocks.graph import add_annotation
-from blocks.initialization import Constant
-from blocks.filter import VariableFilter, get_application_call
-from blocks.roles import (INPUT, WEIGHT, BIAS, BATCH_NORM_POPULATION_MEAN,
-                          BATCH_NORM_POPULATION_STDEV, BATCH_NORM_OFFSET,
-                          BATCH_NORM_DIVISOR, BATCH_NORM_MINIBATCH_ESTIMATE,
-                          add_role)
-from blocks.utils import (shared_floatx_zeros, shared_floatx,
-                          shared_floatx_nans)
 from picklable_itertools.extras import equizip
+from theano.tensor.nnet import bn
+
+from ..graph import add_annotation
+from ..initialization import Constant
+from ..roles import (WEIGHT, BIAS, BATCH_NORM_POPULATION_MEAN,
+                     BATCH_NORM_POPULATION_STDEV, BATCH_NORM_OFFSET,
+                     BATCH_NORM_DIVISOR, add_role)
+from ..utils import (shared_floatx_zeros, shared_floatx,
+                     shared_floatx_nans)
+from .base import lazy, application
+from .sequences import Sequence, Feedforward, MLP
+from .interfaces import RNGMixin
 
 
 class BatchNormalization(RNGMixin, Feedforward):
-    """Normalizes activations, parameterizes a scale and shift.
+    r"""Normalizes activations, parameterizes a scale and shift.
 
     Parameters
     ----------
@@ -254,83 +254,3 @@ class BatchNormalizedMLP(MLP):
         # which is the input dimension.
         for act, dim in equizip(self.activations, self.dims[1:]):
             act.children[0].input_dim = dim
-
-
-def batch_normalize(computation_graph, epsilon=1e-4):
-    """Activate batch normalization in a graph.
-
-    Parameters
-    ----------
-    computation_graph : instance of :class:`ComputationGraph`
-          The computation graph containing :class:`BatchNormalization`
-          brick applications.
-    epsilon : float, optional
-        The stabilizing constant for the minibatch standard deviation
-        computation. Added to the variance inside the square root, as
-        in the batch normalization paper.
-
-    Returns
-    -------
-    batch_normed_computation_graph : instance of :class:`ComputationGraph`
-          The computation graph, with :class:`BatchNormalization`
-          applications transformed to use minibatch statistics instead
-          of accumulated population statistics.
-
-    Notes
-    -----
-    Assumes the minibatch axis is 0. Other axes are unsupported at
-    this time.
-
-    """
-
-    # Create filters for variables involved in a batch normalization brick
-    # application.
-    def make_variable_filter(role):
-        return VariableFilter(bricks=[BatchNormalization], roles=[role])
-
-    mean_filter, stdev_filter, input_filter = map(make_variable_filter,
-                                                  [BATCH_NORM_OFFSET,
-                                                   BATCH_NORM_DIVISOR, INPUT])
-
-    # Group means, standard deviations, and inputs into dicts indexed by
-    # application call.
-    def get_application_call_dict(variable_filter):
-        return collections.OrderedDict((get_application_call(v), v) for v in
-                                       variable_filter(computation_graph))
-
-    means, stdevs, inputs = map(get_application_call_dict,
-                                [mean_filter, stdev_filter, input_filter])
-
-    assert (set(means.keys()) == set(stdevs.keys()) and
-            set(means.keys()) == set(inputs.keys()))
-    assert set(means.values()).isdisjoint(stdevs.values())
-
-    replacements = []
-    # Perform replacement for each application call.
-    for application_call in means:
-        axes = tuple(i for i, b in enumerate(means[application_call]
-                                             .broadcastable) if b)
-        minibatch_mean = inputs[application_call].mean(axis=axes,
-                                                       keepdims=True)
-        minibatch_mean.name = 'minibatch_offset'
-        # Stabilize in the same way as the batch normalization manuscript.
-        minibatch_std = tensor.sqrt(tensor.var(inputs[application_call],
-                                               axis=axes, keepdims=True)
-                                    + epsilon)
-        minibatch_std.name = 'minibatch_divisor'
-
-        def prepare_replacement(old, new, role, application_call):
-            """Add roles and tags to replaced variables."""
-            add_role(new, BATCH_NORM_MINIBATCH_ESTIMATE)
-            add_role(new, role)
-            add_annotation(new, application_call)
-            add_annotation(new, application_call.application.brick)
-            new.tag.replacement_of = old
-            replacements.append((old, new))
-
-        prepare_replacement(means[application_call], minibatch_mean,
-                            BATCH_NORM_OFFSET, application_call)
-        prepare_replacement(stdevs[application_call], minibatch_std,
-                            BATCH_NORM_DIVISOR, application_call)
-
-    return computation_graph.replace(replacements)
